switched to idents, all still works

+++ src/main/scala/caffeineScript/ir/AST.scala
@@ -11,8 +11,7 @@ case class Instruction(ingredient: Ingredient, quantity: Quantity, verb: Verb) e
 
 case class Ingredient(name: String)
 
-case class Quantity(typename: String, amount: Int)
-// could change amount to double?
+case class Quantity(typename: String, amount: Double)
 // should there be a standard measurement for volume/mass?
 // do the benefits of the canonical representation outweigh the drawbacks?
